# docs/api/generated-swagger.yaml
openapi: 3.0.3
info:
  title: NextDocs API
  description: |-
    Auto-generated API documentation for NextDocs

    Generated on: 2025-12-14T09:37:21.234Z
  version: 1.0.0-20251214
  contact:
    name: NextDocs Team
servers:
  - url: /api
    description: API Server
security:
  - sessionAuth: []
  - apiKeyAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: NextAuth.js session cookie
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: hex
      description: API Key authentication (64-char hex string)
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error description
      required:
        - error
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /webhooks/github:
    post:
      tags:
        - Webhooks
      summary: POST /webhooks/github
      description: Auto-generated documentation for POST /webhooks/github
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {}
...